requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
anyio==4.3.0
arrow==1.3.0
asgiref==3.7.2
attrs==24.3.0
binaryornot==0.4.4
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cookiecutter==2.6.0
Django==4.2.6
djangorestframework==3.14.0
easygui==0.98.3
flatbuffers==24.12.23
flet==0.20.2
flet-core==0.20.2
flet-runtime==0.20.2
frozenlist==1.5.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
humanfriendly==10.0
idna==3.6
imageio==2.36.1
Jinja2==3.1.3
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
oauthlib==3.2.2
onnxruntime==1.20.1
opencv-python-headless==4.11.0.86
packaging==23.2
Pillow==10.1.0
platformdirs==4.3.6
pooch==1.8.2
propcache==0.2.1
protobuf==5.29.3
Pygments==2.17.2
PyJWT==2.10.1
PyMatting==1.1.13
pypng==0.20220715.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
referencing==0.36.1
rembg==2.0.61
repath==0.9.0
requests==2.31.0
rich==13.7.1
rpds-py==0.22.3
scikit-image==0.25.0
scipy==1.15.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
sympy==1.13.3
text-unidecode==1.3
tifffile==2025.1.10
tqdm==4.67.1
twilio==9.4.3
types-python-dateutil==2.8.19.20240106
typing_extensions==4.10.0
tzdata==2023.3
urllib3==2.2.1
watchdog==3.0.0
websocket-client==1.7.0
websockets==11.0.3
whitenoise==6.6.0
yarl==1.18.3
